backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "nfactorial-code-agent-backend"
version = "0.1.0"
description = "Backend for the nfactorial code agent (FastAPI)"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv==1.0.0",
  "pydantic",
  "fastapi~=0.117.0",
  "uvicorn",
  "openai-agents[litellm]",
  "PyJWT==2.9.0",
  "vercel-sandbox==0.0.1",
  "vercel-sdk==0.0.5",
  "httpx",
  "httpx-sse",
]

[project.urls]
homepage = "https://example.com"
repository = "https://example.com/repo"


[tool.hatch.build.targets.wheel]
# Include the src package so Hatch can build the wheel after refactor
only-include = [
  "src",
]


